Stripped implementation from wsendpoint.hpp
